+++ src/components/button/__next__/button.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from "@storybook/react";
-import React, { useState } from "react";
-import Button, { ButtonProps } from "./button.component";
+import React, { useRef, useState } from "react";
+import Button, { ButtonHandle, ButtonProps } from "./button.component";
 import Box from "../../box";
 import Icon from "../../icon";
 import { Loader } from "../../loader/__next__/loader.component";
@@ -228,3 +228,22 @@ export const ButtonAsALink: Story = () => {
   );
 };
 ButtonAsALink.storyName = "As a Link";
+
+export const ProgrammaticFocus: Story = () => {
+  const buttonRef = useRef<ButtonHandle>(null);
+
+  return (
+    <Box display="flex" gap={2}>
+      <Button ref={buttonRef} variantType="primary">
+        Button to Focus
+      </Button>
+      <Button
+        variantType="secondary"
+        onClick={() => buttonRef.current?.focusButton()}
+      >
+        Focus other button
+      </Button>
+    </Box>
+  );
+};
+ProgrammaticFocus.storyName = "Programmatic Focus";