+++ src/components/button/__next__/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./button.component";
 export type { ButtonProps } from "./button.component";
+export type { ButtonHandle } from "./button.component";